backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.1
python-decouple==3.8
django-filter==23.5
Pillow==10.2.0
openpyxl==3.1.2
reportlab==4.0.9
